+++ core/org.eclipse.cdt.core/utils/org/eclipse/cdt/utils/debug/tools/DebugSym.java
@@ -35,7 +35,7 @@ public class DebugSym implements Comparable {
 		} else if (obj instanceof Long) {
 			Long val = (Long) obj;
 			anotherVal = val.longValue();
-			thisVal = (long) this.addr;
+			thisVal = this.addr;
 		}
 		return (thisVal < anotherVal ? -1 : (thisVal == anotherVal ? 0 : 1));
 	}
